requirements.txt
google-genai
colorama
python-dotenv
Python 3.6